cleaning up new login flow
